Such a minor change spanning so many files...
Ahem. Replaced the npcl() array with an npc() array of NPCInst UDTs. Works fine in QB, and is a heck of a lot more readable now. Not to mention that npc data is stored consecutively in memory, and thus can be cached by the processor...

If this screwed up anyone's patch, I appologize.

TBD: modify custom to use the same thing (right now it doesn't, although I've modified the compiling scripts so that it still links to the NPC loading routines. Go figure.)


git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@580 7d344553-34f0-0310-a9b1-970ce8f1c3a2