Fixed a few bugs introduced to inputstring by the last couple of revisions.
Guys, PLEASE don't use random indentation, and avoid tabs. And try and keep a consistent style.

And while I was at it, I totally revamped the considerably ugly inputscript for user friendliness (and dev unfriendliness). For instance, ESC no long brings up the menu, the string is now shown if not visible (and returned to its previous visibility afterwards), strings are trimmed to maxlen, etc

Also, I got rid of the showvalue, that looked ugly.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@605 7d344553-34f0-0310-a9b1-970ce8f1c3a2