new "virtual scancodes" for key is pressed:
* joy:button 1 (through 16)
* joy:x left (and right)
* joy:y up (and down)

These allow one to use the key is pressed command to read the joy stick in the same manner that you read the keyboard. This is a good thing, especially for key mapping.

In addition, I've further broken compatability with QB by including gfx.bi in yetmore.bas directly, so I didn't have to write allmodex wrappers for the I/O functions.

Finally, documentation will be coming shortly.


git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@638 7d344553-34f0-0310-a9b1-970ce8f1c3a2