Unlump modified to work with FBC.
However, something seems odd... I copied in unlumpfile() (amongst a number of functions, so I didn't have to compile allmodex.bas with it), but it didn't work out of the box. It was doing some weird crap, which may be related to a change I made on line 605 and 606, to deference the pointer rather than try to load data *on top* of the pointer (changed bufr -> *bufr).

When I did that, the odd behaviour I noticed went away (specifically that of size going insane (probably because, in memory, it's located right next to, or near, bufr)). However, does the version in allmodex.bas need this fix? Unlike this instance, it seems to work ok... but, who knows...?


git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@686 7d344553-34f0-0310-a9b1-970ce8f1c3a2