A couple of related changes:
First, a "timer" feature. A timer is a count down of sorts, at the end of which something happens (either game over, or a plotscript). It has the ability to run during a battle, and if it runs out during such a time, it will either terminate the battle, or politely wait until it's finished (depending on a flag). More details to be added to the wiki soon.

Next, I updated the plotstrings to use a UDT. It looks a lot cleaner now, and when we eventually get around to having auto-strings, this should make it easier.

Finally, I updated the plotscripting dictionary to add support for <note>, <warning> and <danger> tags, which will box and highlight text with an icon, to indicate things the plotscripter should be aware of.

(I know that you hate it when we group lots of changes together into one commit. But, I couldn't help it this time... all the changes were made together...)


git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@936 7d344553-34f0-0310-a9b1-970ce8f1c3a2