fix bug 279 (linux crash with FB 0.16b)
There was a problem with the loadattackdata sub. It was still trying to
load data even if the attack data getbinsize was zero, which resulted
in a temporary array being allocated with (-1) dimentions. Surprisingly
this did not result in an error, but subsequent attempts to write to
this array caused other parts of memory to be overwritten.

The fact that this did not crash in Windows or in Linux with 0.15b was
sheer luck.


git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@957 7d344553-34f0-0310-a9b1-970ce8f1c3a2