New music backend
Ever since I wrote music_native, I've been unhappy with it. It was inefficient, and due to the fact that I didn't (and still don't) know how to do midi timing properly, it was horribly laggy.

Now, Windows has a couple interfaces for playing midi data, from the high-level MCI interface (set it and forget it), to the low-level raw interface (used currently). I opted for the latter, so that we could have cool things like loop points and whatnot.

However, Windows has a third interface, a mid-level "stream" interface. Basically, you break up your song into chunks, and feed them to windows, and it can time and play them for you. The idea is to get the best of both worlds: Control over how it's played, and good timing.

Initially, I wanted to use this, but I was put off by the extra work needed to set it up. However, last week, I decided that enough was enough. Thus, music_native2.

Right from the get go, you'll be able to hear the difference. I've even retained the ability to loop using the RPG Maker style loop points. However, I've decided to drop the extended sysex form, as it was overly complex to begin with, and would probably make this code as bad as what it succeeds.

It's not perfect yet, however. A few songs I've tested don't play properly, but I'm not sure why yet. Further, there's a bit of debugging code left in which, while light weight in terms of speed impact, will likely be very annoying (I'll leave it to everyone else to see what it is, as it in itself is cool.).

So, to compile with music_native2, you can either edit make*.bat to tack a "2" on to the music back end, or use make*_gxf_music.bat directly:

makegame_gfx_music fb native2


git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@960 7d344553-34f0-0310-a9b1-970ce8f1c3a2