Custom now inserts default hand points for old games.
These points are bogus in terms of actually being on the hero's hands, BUT they reproduce old weapon behaviour.

I've tried to minimise the damage to games that already use hand points by checking to see if any are set, but I only check the hero's hand points, not the weapon handle points. This is because I only clobber the hand points. I suspect that the situation of only using weapon handle points for alignment is rare anyway, so I won't lose any sleep over it.


git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@992 7d344553-34f0-0310-a9b1-970ce8f1c3a2