Script loading rewrite: Replaced the old script buffer with a script caching scheme - overflows are obsolete. This also hugely speeds up the interpreter.
Scripts are kept loaded in memory after they finish, until the sum of all loaded scripts exceeds scriptmemMax (in const.bi), at which point scripts are picked to be freed, based upon size, usage and time since last use. Scripts that are currently in use by dormant script instances can be freed too - they're reloaded when needed. So scriptmemMax imposes only a max script size (I choose 256KB - the old (32bit) limit was 16KB).

Created a ScriptData udts.bi for loaded scripts.

I fixed a problem with string tables - the offset was 16bit, limiting max size of scripts with strings to 32KB. I incremented the script format to 2 because of this.

Note: size of script() is arbitrary, I just decided to make it 128.

-added loadscript, freescripts, killallscripts, resetinterpreter

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@1008 7d344553-34f0-0310-a9b1-970ce8f1c3a2