Audwrap/music_native works on linux
There are secrets in this world, great mysteries that cannot, nay, must not be revealed to mortals, lest the great technogods be forced to cleanse the earth of our presence for fear of their position in the heavens...

One of these mysteries is how I managed to get music_native to compile on linux. I fixed up audwrap, and managed to make it link to the OHR. I know not how, but I hope that it works for other people.

You should grab the source for audiere, compile and install it. That requires libogg, libvorbis and libvorbisfile which I assume everyone already has. If not, pop over to http://vorbis.com to get them (I, again, grabbed the source and compiled them; libvorbisfile comes with libvorbis, it's not a seprate package).

Other than that, just compile with ./makegame.sh fb native and ./makeedit.sh fb native, and you'll be fine.

NOTE: Midi does not work yet. I wasted all weekend just geting the stub to compile. :(



git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@1078 7d344553-34f0-0310-a9b1-970ce8f1c3a2