Sound effecting pausing was broken in music_native. It would start over when you tried to play it again.
AudSetRepeat resets the stream if it is not playing (Audiere, not audwarp, problem)
It's possible to change the looping behaviour of a sound effect when resuming it.
It would also be possible to change the looping behaviour of a currently playing sound effect, but I'm not sure music_sdl could match that ability.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@1298 7d344553-34f0-0310-a9b1-970ce8f1c3a2