New development too: check.bat
running check.bat will allow you to check to see if a given source file is syntactically correct. It is similar to running "fbc file.bas", but will not produce any of the linking errors.

To run it, call "check <file>". <file> is the name of the source module, WITHOUT THE EXTENSION. This is important, as the windows command line parser is stupid, and would not be able to delete the residual .o file that is produced as a byproduct of compiling the file.



git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@1588 7d344553-34f0-0310-a9b1-970ce8f1c3a2