Added a "call" command to the front of each "fbc" line in the compilation batch files.
The reason I added it is because on my computer, "fbc" is actually a batch file. I keep several versions around, for compatability testing, in folders like "freebasic.17", "freebasic.18", etc, and the batch file redirects to the correct folder with the "FBVER" environment variable. Works really nice.

I know that I'm the only such person, but the call is benign when used on real executables, so I am just reclaiming the ability to type "svn ci" to commit everything.


git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@1827 7d344553-34f0-0310-a9b1-970ce8f1c3a2