Two things of note:
1. Implemented a new slice type, "Text". Does what it says on the tin. To be done: add word-wrapping, etc.

2. Implemented attachment. There are three types:

slParent - the X and Y of the child is added to the ScreenX and ScreenY of the parent. So, the child moves with the parent
slScreen - the X and Y are relative to the screen. In other words, moving other slices won't affect this one.
slSlice  - the X and Y are relative to an arbitrary other slice. This can be used to attach to e.g. NPCs

Hmm, since slParent is a special case of slSlice, I may just combine them...



git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@2325 7d344553-34f0-0310-a9b1-970ce8f1c3a2