Added setsliceextra, getsliceextra, and increased the number of NPC extra data fields to 3 (so menu items, slices, and NPCs all have 3)
I changed the values of the constants extra1 and extra2 (and added extra3) and the way that the extra data field is calculated in [set]NPCextra. I could do this because extra data isn't saved in save games, and because there is now room for off-by-1 mistakes. I'm assuming that noone is relying on NPCextra[n] being identified to NPCextra[n+2].

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@2589 7d344553-34f0-0310-a9b1-970ce8f1c3a2