Lots of progress on an integrated help system. Press F1 almost anywhere in custom.
I changed the Fkey shortcuts in the map editor so they are all 1 higher now. Don't hate me!

Currently this only allows a single-screen page of help. I need to implement scrolling for longer help 
screens.

Generally, this help is for general menu/editor info and magical-keypress summaries. It does not attempt 
to solve the problem of providing help descriptions for each data element in an editor (that will be 
better served later with the Editor Editor)

Adding help files for a few screens now, lots more to come later after I have finished scrolling.


git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@2645 7d344553-34f0-0310-a9b1-970ce8f1c3a2