If you use "Erase Map Data" on the very last map, the map data files will be completely removed from your RPG file. All other maps still keep the old deletion behaviour of blanking data and shrinking the map to a small size.
Note that if you use this feature to delete maps from a pre-unlumped
RPG file, you will also need to manually remove the map lumps yourself.

Fixed the fatalerror command to reset the palette before displaying
the error message. This was important for errors that happen during a
fade-to-black (for example, walking through a door that leads to a map
that you have deleted)


git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@2688 7d344553-34f0-0310-a9b1-970ce8f1c3a2