Fix Bug 756 - Holding ENTER or SPACE while dismounting a vehicle can result the hero getting stuck
This was caused by a double-dismount triggered if the dismount key was being pressed on the
very last tick of the dismount sequence, cusing the hero to think he needed to initialize
a dismount when the dismount had in fact just completed.

This was easily fixed with a single line check.
Gosh the code is easier to maintain when it is cleaned up a bit :)


git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@2847 7d344553-34f0-0310-a9b1-970ce8f1c3a2