Wrote a proper lexer for HSpeak - nearly: it still produces strings rather than lexemes since the rest of HSpeak wouldn't know what to do with them.
", $, and = are no longer valid characters in identifiers. Unfortunately this was not warned about, although I could have told you this was going to happen. HSpeak disallows any use of $, =, or strings outside of plotstrings and script arguments.

This is required for script error line number reporting for a very round-about reason. More importantly, it is a big step towards HS4.

As a bonus, the new lexer is 4 times faster than the old one.


git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@3466 7d344553-34f0-0310-a9b1-970ce8f1c3a2