Added lastChild to Reload.Node, to speed up AddChild.
This reduces time taken to 'chug' (not including optimising) that 64MB XML from 75 -> 6 sec for me. (Total time to load and convert to RELOAD: 90 -> 22 sec.)

I also fixed AddSiblingBefore, which would have corrupted a reload tree if it had been used to insert before the first child (but isn't used anywhere).

And I disabled the RELOAD_INTERNAL shield when compiling with -g, because it's EXTREMELY annoying.

git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@3575 7d344553-34f0-0310-a9b1-970ce8f1c3a2