Rewrote script usage and broken triggers menus and trigger autofix, removing triple code duplication.
It now takes 3.5 seconds instead of 500 seconds to build the script usage menu in Powerstick Man XE on this laptop! (That was the slowest way to sort a list that I have ever seen in serious use)

Also, fixed a couple bugs, including crashing when there were no scripts imported.

Also, the helpfile lied: previously the script usage menu did not sort scripts.

git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@3626 7d344553-34f0-0310-a9b1-970ce8f1c3a2