Support for variable record number and length .N## lumps, including variable number of NPC definitions.
I have not actually changed the .N## record length yet.
You can add or remove NPC definitions and delete all copies of an NPC on the map from the NPC editor; the default number of NPCs is just 1 (and old games no longer automatically jump to 100 NPC definitions).
The meaning of max_npc_defs changed. I decided to set the limit to 500 for now, we should probably raise it when more NPC instances are possible.

git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@3628 7d344553-34f0-0310-a9b1-970ce8f1c3a2