Windows builds were completely broken when using FB 0.21 or later.
Turns out that __FB_UNIX__ is *always* defined, to either true or false, unlike all the other __FB_*__ defines. Good work, FB devs. Actually I'd already complained about this on the forum... to the guy who wrote the patch for this when he was asking for feedback. He fixed all the other problems I pointed out :(

git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@3847 7d344553-34f0-0310-a9b1-970ce8f1c3a2