gfx_directx: Preparing mouse to handle cursor visibility _and_ cursor clipping requests from both the engine and the backend.
A currently active example for visibility requests: the engine requires the mouse for drawing on a map tile, providing its own cursor. That is the engine request. When the user clicks and holds, the mouse cannot leave the window until the user releases the mouse. That is the backend request.

git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@3883 7d344553-34f0-0310-a9b1-970ce8f1c3a2