Added CapsLock check to strgrabber. It won't work, though.
I really can't remember how to write in basic anymore. :( In the clearly marked segment, inappropriately marked hexadecimal numbers, bitwise operations, and probably branch logic syntax sits. The & is the familiar c-style ANDing, | is ORing, and HxFE is the c equivalent 0xFE. I apologize for my gross failure in this commit, though it won't be hard to fix.

git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@3953 7d344553-34f0-0310-a9b1-970ce8f1c3a2