Add scroll tool to the sprite editor too, replacing shift-scrolling (follows r3957).
It repulses me to be working on code which should be REWRITTEN FROM SCRATCH. Well, some bits of the sprite editor might be salvageable, unlike the tile editor which is 99.5% horrible. However, my impression of the sprite editor after this is that it's really bad code cleaned up to look nicer, but still fundamentally bad.

git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@3962 7d344553-34f0-0310-a9b1-970ce8f1c3a2