revision 4148 broke pasting in the sprite editor browser. Whenever you paste a sprite, intgrabber_with_addset() jumps you back to the first row so your paste always goes in the first row no matter what.
This revision adds a use_clipboard argument to intgrabber() that can optionally
skip the copy/paste code. it defaults to YES, but when intgrabber_with_addset()
calls intgrabber() it is forced to NO... 

...except for mysterious and frustrating reasons, intgrabber() always behaves
as if you passed YES to the use_clipboard argument, regardless of what you
actually pass. Grrrrrr!

So this checkin does NOT fix the sprite pasting problem, but it SHOULD :(


git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@4163 7d344553-34f0-0310-a9b1-970ce8f1c3a2