Add Unix+Windows (untested) file locking functions with timeouts, because I didn't like FB's (which are buggy anyway)
Note the disparity between Unix and Windows versions: on Unix the locking is
advisory (meaning it only blocks other processes from locking), while on Windows
it blocks everything.

For our purposes, the Unix approach is greatly preferred. File locking in
Windows frequently seems to lead to problems, and I'm worried about deadlocks.
Maybe best not to use the Windows functions.

git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@4652 7d344553-34f0-0310-a9b1-970ce8f1c3a2