Fix harm tile bug via a gen pref bitset: when the caterpillar party is disabled, only the leader is harmed. The "harm tiles harm non-caterpillar heroes" bitset defaults to on in new and ancient games.
This bug was introduced in espereble, July 2003. Unfortunately, there's no accurate way to determine whether an RPG was made before espereble. The best I could manage is testing the RPG format version to enable the bugfix for games made before March 2001!

Hey guess what, we still don't store the exact version that an RPG file was made with >:(

git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@4797 7d344553-34f0-0310-a9b1-970ce8f1c3a2