Add "textbox line", replacing "string from textbox", which was buggy, and adding an option specifying whether to strip whitespace:
Not only did stringfromtextbox use NOT when it shouldn't have for the 'expand' argument, but the documentation for this argument was also the wrong way around. It expanded when the argument wasn't -1.

I used scanscripts.py to check that apparently noone has ever passed -1 for this argument, and all or nearly all games were made after 'NOT' changed to bitwise.

git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@4858 7d344553-34f0-0310-a9b1-970ce8f1c3a2