New default and "Basic Unicode" fonts.
"Basic Unicode" is the first 256 code points (which has lots of non-printables),
and the new default is Basic Unicode plus most of the icons in the old default.

Luckily, Basic Unicode includes nearly everything on any non-Cyrillic or -Greek
European keyboard -- awesome!

This new font is totally full -- I had to remove some of the icons which I
couldn't make out/found least useful from the old font. I based the font on the
"ASCII" font (which is not actually ASCII, but rather some IBM extended ASCII
codepage), but ended up redrawing most of it. Some of my characters are ugly;
8x8 is too small for accented capitals.

A lot of the first 256 unicode code points -- the 3/4 icon, say -- aren't very
useful and could potentially be replaced.

git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@4963 7d344553-34f0-0310-a9b1-970ce8f1c3a2