Mac build script now recompiles hspeak, and includes it in the correct location in the OHRRPGCE-Custom.app bundle.
Reorganize and simplify the layout of files in the .dmg to make it a little less confusing.

Create a new README-mac.txt with mac-specific install information, and include it instead
of README-game.txt and README-custom.txt

Put unlump and relump, and the non-bundled hspeak into a seprate zip, since they don't need to be in the .dmg for typical mac users.


git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@5221 7d344553-34f0-0310-a9b1-970ce8f1c3a2