Mac: fix the arrow key problem when browsing spritesets in the sprite editor. I don't understand why this was happening, nor why it was Mac-only, this is just a workaround.
The sprite set browser was the only place that passed scUp and scDown to intgrabber
(by way of intgrabber_with_addset) and apparently that was triggering the problem.
I changed the intgrabber keys to period and comma which didn't have any meaning on
that screen before. You can press period and comma to reproduce the underlying bug.


git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@5269 7d344553-34f0-0310-a9b1-970ce8f1c3a2