Fix useinn when called directly from a textbox (instead of from a shop) It was restoring the party even if you refused to pay.
The old useinn gosub code ised the "inn" variable both as a menu cursor
and as the return value. When it was cleaned up as a proper function
it used its return value for when it was called by a shop, but it continued to
also use the old inn variable even though it was being passed byval

Fixed by removing the inn variable as an argument. It is now only an internal menu
cursor variable, and fixed the calling code to only consider the return value.


git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@5376 7d344553-34f0-0310-a9b1-970ce8f1c3a2