If the vehicle shadow colour was 0 then no shadow was drawn.
Rather than fix this properly, I worked around this by changing uilook(uiShadow) to the nearest non-zero match in upgrade()

A proper fix would be to allow colour 0 ellipses (which would be possible using Frames with a mask). However that would need a gen. pref. bitset to enable. Too messy.

git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@5379 7d344553-34f0-0310-a9b1-970ce8f1c3a2