Fix bug preventing lumps from being lumped if there are any lumps ending in .tmp (main part of Bug 977 fix)
The ones excluded are the final ones in the listing returned by findfiles. On
Windows, this seems to be sorted in alphabetical order, so UICOLORS.BIN would
usually be the first file to get the chop.

Bug was introduced in r3897 "findfiles now returns its results in a array instead of a file..."
(2010-10-17, included since Zenzizenzic)

git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@5529 7d344553-34f0-0310-a9b1-970ce8f1c3a2