Fix Mac port; was broken by work for -gen gcc support.
I mistakenly thought __FB_GCC__ is only defined when compiling with -gen gcc
but actually it's always defined to either 0 or 1.

This caused the setjmp GOSUB hack to be used (on all platforms) instead of the
call/ret hack which we had been using. I hadn't realised it before, but the
setjmp GOSUB hack doesn't work on Mac because the tricky saving of the esp
register which the Mac port of FB's stack alignment hack uses is interfered
with by setjmp's register saving/loading.

This patch goes back to using the call/ret hack as intended.

I also made the same mistake with __FB_GCC__ in vector.bi, so the alternative
function declarations were always being used. And it turns out that the 'main'
declarations didn't work with a recent testcase, so I fixed that.

git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@5728 7d344553-34f0-0310-a9b1-970ce8f1c3a2