Add mark+clone to map editor!
You can't yet copy paste between maps.

The implementation uses vectors of MapEditUndoTiles, like the Undo
feature, but uses separate create_changelist and apply_changelist
functions for handlign them.

In addition, there is a new system to allow previewing map edits (used
for previewing cloning and also the Draw tool in Tilemap mode) which
also uses the undo system; there is a secondary_undo_buffer for
this. Took me days to decide upon, but it's very simple and powerful.

git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@5744 7d344553-34f0-0310-a9b1-970ce8f1c3a2