Delay map ambient music by a tick, to give scripts a chance to override it without it playing for a moment.
"current song" will pretend that any pending music change has already happened,
to avoid confusion.

This fixes a music glitch at the beginning of VoM and should fix a number of
glitches in SoJ and possibly Powerstick Man.

I previously added the "suspend map music" command to deal with the same
problem (probably for Powerstick Man?) but this is much cleaner.

git-svn-id: svn://gilgamesh.HamsterRepublic.com/ohrrpgce/wip@5963 7d344553-34f0-0310-a9b1-970ce8f1c3a2