Increase fade in delay to 1 tick after loading a save during a game, or healing or battle triggered by a textbox
Previously it was not possible for the load game plotscript to do
anything before the screen faded in (from a load triggered with
wantloadgame rather than the pos-title screen load menu).
I also added non-zero delays to the other fades which didn't have them.

For back-compat, the newly extended fades can be cancelled by a
fadescreenout command. This is so that if a script used to call
fadescreenout after one of these fades happened, but it now happens
before the scheduled fade, then the fade is cancelled rather than
overriding the scripted fade.

This also has the nice advantage that by simply putting "fadescreenout"
at the top of your load game plotscript, you cancel the automatic
fadein and can take as long as you want to set things up, rather
than the implicit waits in teleporttomap, etc, ruining your plans.
So I think it would be great if *all* queued fades were overridable
with fadescreenout, but that would cause other problems if people
do a fadescreenout and rely on the fade in to happen automatically.
Add a backcompat bit maybe?

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@6457 7d344553-34f0-0310-a9b1-970ce8f1c3a2