Fix bug 763: walkabouts flicker when crossing map edge
Unfortunately there is a circular dependency betwen updating the
camera position, and updating walkabout slice positions, so my
simple solution is to update walkabout positions twice.

I think the best solution is to have the slice system handle
wrapped drawing of slices over the map edge rather than updating
walkabout slice positions (that is, get rid of framewalkabout),
as this fixes the display of all slices parented to map layers, not
just walkabouts. It would even be possible to draw the same slice
multiple times, when the map size is <= the screen size!

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@6471 7d344553-34f0-0310-a9b1-970ce8f1c3a2