plotdict: Fix another XSLT problem causing links to some commands to disappear in Firefox
It appears that when a command had a <reference> that appeared before the
definition, Firefox tried looking up 'shortname' in that <reference> instead
of the <command>, because both had the same id. I don't know whether this is
a bug in Firefox, but the duped 'id' was in xml, not html, and xmllint
and Chromium didn't complain, so I think it is.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@7277 7d344553-34f0-0310-a9b1-970ce8f1c3a2