Change the way start position of projectile attacks is calculated.
The center of the attack sprite is now aligned to center of left or
right edge of attacker sprite, rather than appearing +/- 50 pixels
left or right of the top-left corner of the attacker.
Also, the real rather than initial position of the attacker is now
used.

Probably noone will ever notice the difference.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@7518 7d344553-34f0-0310-a9b1-970ce8f1c3a2