Fix broken parallax and scrolling water in 'Char's Lovely Home'
This adds a backcompat bit "Wrap map layers over edge of Crop maps"
and a fixbit fixWrapCroppedMapsBit, and is the first game upgrade to
make use of fine-grained version info.

In r6473 (earler 2013 after Beelzebufo) I changed the behaviour of
cropping map layers to truely crop rather than wrap. This only matters
if you view a position off the edge of a map layer, either by moving
them about or if the map is smaller than the screen. The latter is why
I noticed and changed that in r6473.

Giz demonstrated the effects in Char's Lovely Home:
http://www.slimesalad.com/forum/viewtopic.php?t=6777

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@7668 7d344553-34f0-0310-a9b1-970ce8f1c3a2