scons: add asan=1 option for AddressSanitizer, an alternative to valgrind which is far faster
Unfortunately it doesn't seem to work with gcc 4.9.3, tripping
with a false positive 'stack-buffer-overflow' in array_new and then a
failed assert. Looks like an Asan bug -- it's not meant to have false positives.

I don't know if it works with other GCC versions or with clang; for
some reason my clang install is missing asan support.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@7679 7d344553-34f0-0310-a9b1-970ce8f1c3a2