email_save_to_developer, only works on Android. Requires 770bfd5 of sdl-android
This emails an rsav file and g_debug.txt. Not sure it actually works; it worked
until I modified the java function to accept more than one file, and then gmail
started crashing and the usb link to my phone stopped working.

Added an option to the F8 debug menu to email the save, but of course you can't
press F8 on a phone, and the option doesn't appear on other platforms.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@7703 7d344553-34f0-0310-a9b1-970ce8f1c3a2