Also include slice editor help files with Game (Mac only).
I opted to special case them rather than rename them to share_*.txt, which
is the naming scheme used for all other shared help files.
Renamed a help file with a typo'ed name.

We really need a central place to record all the data files that need to be
included with Custom and with Game. Right now, help files are only included
with Game on Mac, while we duplicate the list of data files that Custom
needs in half a dozen different places!

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@7750 7d344553-34f0-0310-a9b1-970ce8f1c3a2