Recompile rcedit.exe without depending on mvscp100.dll
Didn't work in Win XP, and according to an issue filed on github, rcedit.exe
didn't work in Win 10 either. Actually I suspect that it didn't work on any
fresh install of Windows.

What a nightmare. rcedit depends on a single function from the ATL library,
which is not part of (my version of) VS Express; instead it's part of the
Windows Driver Development Kit, which was a 620MB download. Also took me forever
to figure out how to fix the rcedit build configuration.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@7823 7d344553-34f0-0310-a9b1-970ce8f1c3a2