Replace undocumented "email saved game" with "send email", and document it.
I decided if I was going to explain it in detail in IRC I might as well
document it, so I might as well make it more flexible.
send email takes 3 args: which save slot if any to attach (it no longer saves
automatically), and subject and body (both have defaults).

Mentioned the change in IMPORTANT-nightly: that thing hasn't been used in over
3 years!

Avoid a debug (instead of debuginfo) if distrib.reld is missing.

Added a couple scripts to test.rpg "testing scripts" menu for emailing.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@7835 7d344553-34f0-0310-a9b1-970ce8f1c3a2