sliceedit: Fix F6 permanently moving the in-game Root slice.
The root slice now gets reparented to a temporary slice which gets moved
around instead, just like Custom.
This means the chequer background now also shows in-game (but usually it's
hidden by map layers)

The 'reposition_slice' arg to slice_editor() was unused so I replaced it.
(Ctrl+F used to shift the slice to 0,0, which I didn't find helpful)

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@7836 7d344553-34f0-0310-a9b1-970ce8f1c3a2