Change the documentation of Release/Debug modes, and mention it in whatsnew
(Sorry if it seems I'm talking in James' voice, but he implemented it)

The documentation used to recommend using Debug mode for new games.
The whole point of Debug mode was to allow us to add new warning and error
messages without breaking existing games! Therefore people should never release
a game with debug mode on.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@7843 7d344553-34f0-0310-a9b1-970ce8f1c3a2