Rewrite F8 Debug Menu. Now includes all debug keys!
In fact, I moved the implementation of all debug keys into the menu,
so that there is no duplication of the implementations.
I did this in a tricky way to allow putting the implementation, combo key
definition, and debug menu caption all together, and make all of them optional.
FB may not be Python, but programming is all about creating your own 'language'
for describing solutions!

Copied the help page from the wiki.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@7853 7d344553-34f0-0310-a9b1-970ce8f1c3a2