multichoice() now makes use of the top half of the screen when the menu is long.
Basically, the box+menu gets centred rather than first centering the box and
putting the menu beneath it.

This was absolutely horrible to get right; I should have converted it to slices
instead.

As a bonus, restore vpage when leaving multichoice, so that the screenshot
option in the F8 debug menu works as expected.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@7854 7d344553-34f0-0310-a9b1-970ce8f1c3a2