Add "Fullscreen" and "Windowed" special menu items.
Each is disabled if already fullscreen/windowed, and is always disabled on
mobile/console. I also disabled them when using gfx_fb on Linux, because
some pretty horrible stuff happens to me there. gfx_sdl on Linux can also
be unpleasant if the window size is small so that SDL tries to switch to a
smaller screen resolution. Otherwise it is fine, so I left it enabled (and
gfx_sdl can be changed to avoid doing that).

Added the new options to the default menu, and updated ohrrpgce.new to match.
I also moved Volume above Quit so that it's:
 Fullscreen/Windowed
 Volume
 Quit
Although that looks a bit alien to me, but seemed to make more sense.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@7874 7d344553-34f0-0310-a9b1-970ce8f1c3a2