Delete store_from_page, load_to_page by rewriting spriteedit_load/save_what_you_see
Sprite editor saving and loading code now converts a placer() buffer to an
integer buffer for loadrecord/storerecord instead of using pointing
store_from_page/load_to_page to a part of a video page.
This is two steps forward and one step back, because the placer() format is
still a nasty way to store a sprite; the code I wrote here will soon be thrown
out.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@7938 7d344553-34f0-0310-a9b1-970ce8f1c3a2