Full FB 0.24+ support: Replace OPEN with OPENFILE nearly everywhere.
Live-previewing previously largely didn't work in FB 0.24+. FB 0.23
and earlier still work.

I left a number of OPENs where the file was clearly not a lump,
e.g. files created by distribmenu, but wasn't consistent about that.
It might even be nice to use OPENFILE when not necessary just for
the ability to log every call to OPEN.
It's also possible to use OPENFILE in utilities (indeed, necessary
in bam2mid) but mostly it is not used in them.

Also removed a now-unneeded call to send_lump_modified_msg() from
Buffered_close.

misc/fbutils.el contains an Emacs script I wrote to do the translation,
so it can be easily repeated.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@8000 7d344553-34f0-0310-a9b1-970ce8f1c3a2