scons/mac: add -mmacosx-min-version=10.4 even if no SDK is linked
I didn't actually see a problem as a result of this being missing, but
it probably would have prevented running on older OSX when not using an
SDK. I do not believe that using a Mac SDK is actually necessary now.

Also, I noticed that linkgcc=0 is broken again on Mac, ironically because
I fixed fbc to process #inclib on Mac, which results in -lSDL, when
we want to link to the SDL framework instead.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@8033 7d344553-34f0-0310-a9b1-970ce8f1c3a2