mac/gfx_sdl: work around SDL bug: SDL_WarpMouse failed to hide the mouse cursor
This only occurred when the mouse cursor was off the window and got
warped onto the window by a call to gfx_setmouse while the mouse
cursor is meant to be hidden, such as in the tile picker when arrow
keys are pressed.

There is definitely an SDL bug here, because I had to call SDL_ShowCursor twice
to work around it, indicating SDL has lost track of the cursor state.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@8036 7d344553-34f0-0310-a9b1-970ce8f1c3a2