Mouse cursor visibility now has 3 states: hidden, visible, default (mouse not in use) Implemented in gfx_alleg/fb/sdl. gfx_alleg not tested.
'default' makes the cursor hidden in fullscreen and visible when windowed.
However gfx_fb loses track of whether it's running fullscreen when alt+enter
is used, so gfx_fb instead defaults to visible in Custom and hidden in Game.

Added defaultmousecursor, getcursorvisibility, setcursorvisibility.

The "show mouse cursor" command now works in fullscreen, as does the slice
editor's use of showmousecursor.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@8045 7d344553-34f0-0310-a9b1-970ce8f1c3a2