gfx_directx: Rewrote cursor visibility handling to support io_setmousevisibility(cursorDefault)
If an old copy of gfx_directx.dll is used with a new game/custom then
cursorDefault acts as cursorHidden. There's no problem if a new copy of
gfx_directx.dll is used with an old game/custom (though I didn't actually test)

Bumped version to 1.19.

I noticed that the use of WM_NCHITTEST to check whether the mouse is on the
client area is dubious according to winapi docs.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@8046 7d344553-34f0-0310-a9b1-970ce8f1c3a2