gfx_directx: Fix bug 2006: slowdown after losing focus while a joystick is plugged in
The slowdown was due to calling refreshEnumeration() every tick after a
device is lost (because we asked for access only while foregrounded), which
is slow as it scans attached hardware.

To prevent the scanning it was sufficient to handle DIERR_INPUTLOST by calling
Acquire() instead of refreshEnumeration(), but I also added handling for
some other errors I saw, and just dropping the device in case of misc Acquire
errors rather than spamming error messages.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@8119 7d344553-34f0-0310-a9b1-970ce8f1c3a2