Get rid of nulzstr, which was unnecessary
Maybe it was necessary back in 2009, but even FB 0.23 accepts both
NULL and @"" as default zstring ptr arguments. The later is a pointer
to a string of length zero. It turns out that @"" is pretty much never
necessary, because unlike C, FB accepts null pointers as valid length
zero strings.