Improve OPENFILE hooking and code cleanup.
Now if you write to write to a write-protected file the write isn't
performed and an error shown immediately, instead of performing it
and showing an error when closing the file.
Likewise copyfile() also checks before instead of after doing the copy.
The hook filter function can now prevent a file from being opened
(relevant because showerror is now used instead of fatal error).
Finally, if you implicitly ask for write access (e.g. don't specify
access) it downgrades to read access if the file is protected.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@8142 7d344553-34f0-0310-a9b1-970ce8f1c3a2