android: Add gdb-commandline-program.sh utility, and rewrite compile-using-toolchain.sh
gdb-commandline-program.sh pushes a binary (not an .apk) to an emulator and
runs it under gdbserver, remotely attaching to it. It's based on ndk-gdb which
does the same thing for .apks.

compile-using-toolchain.sh is now much cleaner, updated to work for both
newer and older NDK releases, and partially works to compile stuff without
needing a standalone toolchain (cross compiling scons without one doesn't
work though)

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@8166 7d344553-34f0-0310-a9b1-970ce8f1c3a2