wstring_to_latin1() can now handle common unicode modifiers, turning them into precomposed chars in Latin1
Specifically for unicode->Latin1 I could have just used wcsrtombs
instead, I guess... that's not supported on Android though.

Also added partially_normalise_unicode() which does the same for codepoints
up to 0x500 (a limit chosen simply to reduce the size of the table
used; the table can be regenerated with more/fewer rewrite rules with
generate_unicode_compose_table.cpp). It is only a partial solution for
simple character compositions, but it'll probably be good enough for a
long time.

git-svn-id: svn://gilgamesh.hamsterrepublic.com/ohrrpgce/wip@8197 7d344553-34f0-0310-a9b1-970ce8f1c3a2