Fix paletted graphics (including text) drawn with wrong colours on Mac.
Actually, this was nothing to do with Mac, but was caused by passing 0 or -1
to C as a bool argument, and clang assuming that a bool is either 0 or 1
(it generated asm that tested "trans == 1" in one place and "trans != 0" in
another, leading to an impossible branch being taken, where the Palette16 ptr
was treated as if it were NULL).

This only happened on Mac because it's clang-specific.