Ensure textsize calculates same size as UpdateTextSlice
This is a bit hacky: each font has a size adjustment, and edged font has
a hardcoded x=-1,y=-1 adjustment.

Although this doesn't perfectly measure the background size it has the advantage
that if you align one text slice to the right of or below another the spacing
will be seamless.