WIP Remove fixDefaultDissolveEnemy, to simplify upgrading and remove danger
fixDefaultDissolveEnemy was added in r1651 (5cdce0fe, 2007-11-25)
because it change the numbering of dissolve types which were added in r1650
(489440a2) a few hours earlier. This was truly excessive, r1650 only added
dissolves in Custom.

Also, the data this fixbit zeroes is already zeroed by upgrade anyway; it used
to contain the enemy death message.

However just because it exists, it increases complexity and results in data loss if loadenemydata+saveenemydata are used in Game.

For example this fixbit was part of the cause of an enemy data loss bug fixed in r9197.