(ABANDONED) Update jo_jpeg by merging changes from jo_jpeg 1.60: U,V subsampling for better compression
I merged in the new version with the changes from the fork we were using (which
adds an optional write callback), and put the result at
https://github.com/rversteegen/jo_jpeg

ABANDONED: it turns out that subsampling saves 10-20% on the size of
photographic images with minor degradation but is absolutely shocking on crisp
line art and pixel art. Therefore it's bad for our use. It could still be
provided as an option.